+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Hexagonal maze - part2".
 - Tests for "Byte pair encoding".
 - Tests for "Find the missing plus signs in addition".
+- Tests for "Rod cutting problem".
 
 ## [1.13.0] - 2022-09-30
 ### Added